refactor(core): split jobs_store god module into routing submodules (sc-8816)
Extract the backend routing/gating/catalog logic out of the 8.8k-line
jobs_store god module into a new `routing/` submodule tree, leaving
jobs_store.rs as the SQLite jobs/workers store plus its SQL-coupled
dispatch. Pure, behavior-preserving code move: no routing decision,
catalog membership, or public API changed.

Layout:
  jobs_store.rs                 SQL store + SQL-coupled dispatch
  jobs_store/routing.rs         declares the 4 submodules
  jobs_store/routing/catalog.rs routed-model/kernel lists + Mac support probes
  jobs_store/routing/mlx.rs     *_mlx_eligible predicates
  jobs_store/routing/candle.rs  *_candle_eligible predicates
  jobs_store/routing/gaps.rs    UnsupportedReason + gap/error builders

jobs_store.rs shrinks 8779 -> 5611 lines (~3187 moved). Moved items are
pub(crate); the already-pub Mac-support/gating surface stays pub and is
re-exported from jobs_store so apps/rust-api and the integration test keep
their `sceneworks_core::jobs_store::{...}` use paths unchanged.

The five bug-prone parallel catalog lists (MLX_ROUTED_MODELS,
CANDLE_ROUTED_MODELS, CANDLE_QUANT_LORA_MODELS, CANDLE_QUANT_MODELS,
CANDLE_LORA_MODELS) were moved together into catalog.rs but NOT
semantically collapsed (their superset/exemption semantics are subtle).
A new membership-parity test snapshots each list's exact contents and
asserts the quant/lora sub-lists are CANDLE_ROUTED_MODELS supersets,
guarding the missed-list-edit bug class (chroma sc-5576, krea sc-7836).
